rename giservice
