+++ apps/passport-server/src/services/generic-issuance/subservices/utils/startGenericIssuanceService.ts
@@ -8,7 +8,7 @@ import { DiscordService } from "../../../discordService";
 import { PagerDutyService } from "../../../pagerDutyService";
 import { PersistentCacheService } from "../../../persistentCacheService";
 import { RollbarService } from "../../../rollbarService";
-import { GenericIssuanceService } from "../../GenericIssuanceService";
+import { GenericIssuanceService } from "../../GenericIssuanceService2";
 
 export async function startGenericIssuanceService(
   context: ApplicationContext,
