+++ apps/passport-server/test/generic-issuance/pipelines/csv/csvPipeline.spec.ts
@@ -13,7 +13,7 @@ import * as MockDate from "mockdate";
 import { stopApplication } from "../../../../src/application";
 import { PipelineDefinitionDB } from "../../../../src/database/queries/pipelineDefinitionDB";
 import { PipelineUserDB } from "../../../../src/database/queries/pipelineUserDB";
-import { GenericIssuanceService } from "../../../../src/services/generic-issuance/GenericIssuanceService";
+import { GenericIssuanceService } from "../../../../src/services/generic-issuance/GenericIssuanceService2";
 import { CSVPipeline } from "../../../../src/services/generic-issuance/pipelines/CSVPipeline/CSVPipeline";
 import { PipelineUser } from "../../../../src/services/generic-issuance/pipelines/types";
 import { Zupass } from "../../../../src/types";
